tests/wpt/web-platform-tests/css/selectors/only-of-type.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Selectors :only-of-type</title>
<link rel="help" href="https://drafts.csswg.org/selectors-4/#the-only-of-type-pseudo">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<!--
  See also child-indexed-pseudo-class.html.
-->
<body>

<div>
  <div id="target1">Whitespace nodes should be ignored.</div>
</div>

<div>
  <div id="target2">A comment node should be ignored.</div>
  <!-- -->
</div>

<div>
  <div id="target3">Non-whitespace text node should be ignored.</div>
  .
</div>

<div>
  <blockquote></blockquote>
  <div id="target4" data-expected="false">There is another child element of a different type.</div>
</div>

<div>
  <div id="target5"></div>
</div>

<script>
for (let i = 1; i <= 4; ++i) {
  let target = document.querySelector(`#target${i}`);
  test(() => {
    if (target.dataset.expected == 'true')
      assert_false(target.matches(':only-of-type'));
    else
      assert_true(target.matches(':only-of-type'));
  }, target.textContent);
}

test(() => {
  const target = document.querySelector('#target5');
  assert_true(target.matches(':only-of-type'));

  const of_different_type =  target.parentNode.appendChild(document.createElement('span'));
  assert_true(of_different_type.matches(':only-of-type'));
  assert_true(target.matches(':only-of-type'));

  const another_of_type = target.parentNode.appendChild(document.createElement('div'));
  assert_false(target.matches(':only-of-type'));
  assert_false(another_of_type.matches(':only-of-type'));

  another_of_type.remove();
  assert_true(target.matches(':only-of-type'));
}, 'Dynamic addition and removal');

</script>
